new scraper

+++ models/Book.php
@@ -33,7 +33,9 @@ class Book extends ModelCommon
 
     public function save_tags($tag_ids) {
         foreach ($this->bookTags as $book_tag) {
-            $book_tag->delete();
+            if($book_tag->tag->type == 0) {
+                $book_tag->delete();
+            }
         }
         $tags_arr = explode(',',$tag_ids);
         foreach ($tags_arr as $tag_id) {
@@ -48,12 +50,22 @@ class Book extends ModelCommon
     }
 
     public function add_tag($tag_name) {
-        $tag = Tag::find()->where(array('name'=>$tag_name))->one();
+        $type = 0;
+        if(strpos($tag_name, 'Author:') !== false) {
+            $tag_name = str_replace('Author:','',$tag_name);
+            $type = 1;
+        }
+        if($tag_name == 'Chưa cập nhật') {
+            return true;
+        }
+        $tag_name = str_replace('Đ','đ',$tag_name);
+        $tag = Tag::find()->where(array('name'=>$tag_name, 'type'=>$type))->one();
         if(empty($tag)) {
             $tag = new Tag();
             $tag->name = $tag_name;
             $tag->slug = generate_key($tag_name);
             $tag->status = Tag::ACTIVE;
+            $tag->type = $type;
             $tag->save();
         }
         if(BookTag::find()->where(array('book_id'=>$this->id, 'tag_id'=>$tag->id))->count() > 0) {
@@ -70,45 +82,11 @@ class Book extends ModelCommon
             $this->image = 'default.jpg';
         }
         $image_dir = \Yii::$app->params['app'].'/web/uploads/books/'.$this->slug;
-
-        if($this->image != 'default.jpg') {
-            if(!file_exists($image_dir.'/'.$this->image)
-            || filesize($image_dir.'/'.$this->image) == 0) {
-                return \Yii::$app->urlManager->createAbsoluteUrl(['/']) . 'uploads/books/default.jpg';
-            }
-        } else {
-            if(empty($this->image_source)) {
-                return \Yii::$app->urlManager->createAbsoluteUrl(['/']) . 'uploads/books/default.jpg';
-            }
-            $image_source = $this->image_source;
-            $array = explode('?', $image_source);
-            $tmp_extension = $array[0];
-            $array = explode('.', $tmp_extension);
-            $extension = trim(strtolower(end($array)));
-            if($extension != 'png') {
-                $extension = 'jpg';
-            }
-            $image = 'cover.'.$extension;
-            $dir_array = explode('/', $image_dir);
-            $tmp_dir = '';
-            foreach ($dir_array as $i => $folder) {
-                $tmp_dir .= '/'.$folder;
-                if($i > 3 && !file_exists($tmp_dir)) {
-                    mkdir($tmp_dir, 0777);
-                }
-            }
-            $image_dir = $image_dir.'/'.$image;
-
-            $ch = curl_init($image_source);
-            $fp = fopen($image_dir, 'wb');
-            curl_setopt($ch, CURLOPT_FILE, $fp);
-            curl_setopt($ch, CURLOPT_HEADER, 0);
-            curl_exec($ch);
-            curl_close($ch);
-            fclose($fp);
-            $this->image = $image;
-            $this->save();
+        if($this->image != 'default.jpg' && !empty($this->slug)
+            && file_exists($image_dir.'/'.$this->image)
+            && filesize($image_dir.'/'.$this->image) > 0 ) {
+            return \Yii::$app->urlManager->createAbsoluteUrl(['/']) . 'uploads/books/' . $this->slug . '/' . $this->image;
         }
-        return \Yii::$app->urlManager->createAbsoluteUrl(['/']) . 'uploads/books/' . $this->slug . '/' . $this->image;
+        return \Yii::$app->urlManager->createAbsoluteUrl(['/']) . 'uploads/books/default.jpg';
     }
 }
